automl_trigger_pubsub/requirements.txt
# requirements.txt
google-cloud-functions
google-cloud-bigquery
google-cloud-aiplatform
pandas
functions-framework
